+++ CHANGELOG.md
@@ -38,6 +38,8 @@ and this project adheres to
 - [FMUDiscipline][gemseo_fmu.disciplines.fmu_discipline.FMUDiscipline] has a new argument, named ``time_name``, to set the name of the time variable (default: ``"time"``).
 - [FMUDiscipline.set_default_execution][gemseo_fmu.disciplines.fmu_discipline.FMUDiscipline.set_default_execution] has a new argument, named ``use_arrays_only``, to pass only NumPy arrays at execution (default: ``False``).
 - The function [plot_time_evolution][gemseo_fmu.utils.plotting.plot_time_evolution] draws the time evolution of a collection of variables.
+- The `time_window` argument of the method [FMUDiscipline.plot][gemseo_fmu.disciplines.fmu_discipline.FMUDiscipline.plot]
+  can also be a float number defining the initial time or a tuple of float numbers defining the initial and final time.
 
 ### Fixed